backend/requirements.txt
Django==5.1.5
psycopg[binary]==3.2.3
python-dotenv==1.0.1
djangorestframework==3.15.2
django-cors-headers==4.6.0
openpyxl==3.1.5
